+++ packages/react-native-matrix-crypto/src/index.ts
@@ -67,6 +67,7 @@ export {
   createCryptoMachine,
   createRecovery,
   decryptEvent,
+  discardScopeKey,
   encryptEvent,
   encryptionSlice,
   exportSecrets,